Moved JobController to API endpoints and changed the URLs in the frontend
