+++ public/js/settings-integration-jellyfin.js
@@ -266,7 +266,7 @@ async function updateSyncOptions() {
 
 async function exportJellyfin() {
     const response = await fetch(
-        '/jobs/schedule/jellyfin-export-history',
+        '/api/jobs/schedule/jellyfin-export-history',
         {'method': 'get'}
     ).catch(function (error) {
         addAlert('alertJellyfinExportHistoryDiv', 'History export could not be scheduled', 'danger')
@@ -296,7 +296,7 @@ async function exportJellyfin() {
 
 async function importJellyfin() {
     const response = await fetch(
-        '/jobs/schedule/jellyfin-import-history',
+        '/api/jobs/schedule/jellyfin-import-history',
         {'method': 'get'}
     ).catch(function (error) {
         addAlert('alertJellyfinImportHistoryDiv', 'History import could not be scheduled', 'danger')
