+++ public/js/settings-integration-plex.js
@@ -226,7 +226,7 @@ document.getElementById('plexServerUrlInput').addEventListener('input', function
 
 async function importPlexWatchlist() {
     const response = await fetch(
-        '/jobs/schedule/plex-watchlist-sync',
+        '/api/jobs/schedule/plex-watchlist-sync',
         {'method': 'get'}
     ).catch(function (error) {
         addAlert('alertPlexWatchlistImportDiv', 'Watchlist import could not be scheduled', 'danger')
