+++ templates/page/settings-integration-letterboxd.html.twig
@@ -34,7 +34,7 @@
                     </ul>
                 </div>
 
-                <form action="/jobs/schedule/letterboxd-diary-sync" method="post" enctype="multipart/form-data">
+                <form action="/api/jobs/schedule/letterboxd-diary-sync" method="post" enctype="multipart/form-data">
                     <h5 style="margin-bottom: 0.5rem">Import History</h5>
                     <p class="text-muted" style="margin-bottom: 0.5rem">Upload the diary.csv file</p>
                     <div class="input-group mb-3">
@@ -62,7 +62,7 @@
 
                 <br>
 
-                <form action="/jobs/schedule/letterboxd-ratings-sync" method="post" enctype="multipart/form-data">
+                <form action="/api/jobs/schedule/letterboxd-ratings-sync" method="post" enctype="multipart/form-data">
                     <h5 style="margin-bottom: 0.5rem">Import Ratings</h5>
                     <p class="text-muted" style="margin-bottom: 0.5rem">Upload the ratings.csv file</p>
                     <div class="input-group mb-3">
