+++ templates/page/settings-integration-trakt.html.twig
@@ -83,7 +83,7 @@
                 <div style="margin-top: 0.9rem">
                     <button type="button" class="btn btn-outline-primary" onclick="showJobModal('trakt_import_history')" style="margin-right: .3rem">Show imports</button>
                     <a class="btn btn-warning {% if coreAccountChangesDisabled == true or traktUserName is null or traktClientId is null %}disabled{% endif %}"
-                       href="/jobs/schedule/trakt-history-sync">Import history</a>
+                       href="/api/jobs/schedule/trakt-history-sync">Import history</a>
 
                     <br>
 
@@ -96,7 +96,7 @@
 
                     <button type="button" class="btn btn-outline-primary" onclick="showJobModal('trakt_import_ratings')" style="margin-right: .3rem;margin-top: 1rem">Show imports</button>
                     <a class="btn btn-warning {% if coreAccountChangesDisabled == true or traktUserName is null or traktClientId is null %}disabled{% endif %}"
-                       href="/jobs/schedule/trakt-ratings-sync" style="margin-top: 1rem">Import ratings</a>
+                       href="/api/jobs/schedule/trakt-ratings-sync" style="margin-top: 1rem">Import ratings</a>
                 </div>
             </div>
         </div>
